// PAGridCustomizerCtrl/types.ts
export type CellEditorOverrides = {
  [dataType in ColumnDataType]?: (defaultProps: CellEditorProps, rendererParams: GetEditorParams)
    => React.ReactElement | null | undefined;
};
// ...
export interface CellEditorProps {
  /** True when cell is secured */
  readonly secured?: boolean;
  /** Raw value of the cell */
  readonly value: unknown;
  /** True when grid is in RTL mode */
  readonly isRTLMode?: boolean;
  /** Grid row height in pixels */
  readonly rowHeight?: number;
  /** True when the cell value is required to be set */
  readonly isRequired?: boolean;
  /** Passing the first character pressed starting the cell editing */
  readonly charPress?: string | null;
  /** Cell column data type */
  readonly columnDataType?: ColumnDataType;
  /** Callback to notify the grid when value of the cell editor is changed */
  readonly onChange: (newValue: unknown) => void;
}

export interface GetEditorParams {
  /** Column definitions for all visible columns in the grid */
  readonly colDefs: ColumnDefinition[];
  /** Cell column index */
  readonly columnIndex: number;
  /** Callback to notify the grid when value of the cell editor is changed */
  readonly onCellValueChanged: (newValue: unknown) => void;
  /** Cell row data */
  readonly rowData?: RowData;
  /** Callback to programmatically stop the cell editing */
  readonly stopEditing: (cancel?: boolean) => void;
}
// ...
export const RECID = '__rec_id';
export interface RowData {
  /** Unique id for the row */
  readonly [RECID]: string;
}

export interface ColumnDefinition {
  /** Field name (should be unique) */
  readonly name: string;
  /** Column display name. Will be shown as column header.If none provided shimmer will be shown in place of header */
  readonly displayName?: string;
  /** The unique ID to give the column. This is optional. If missing, the ID will default to the name.
   * If defined, this ID will be used to identify the column in the API for sorting and filtering. */
  readonly colId?: string;
  /** Data type for the values in the column. */
  readonly dataType: string;
  /** Custom Renderer type to be used for this column when custom renderer override is provided. */
  readonly CustomRendererType?: string;
  /** Width of the column. Auto calculated if not set. */
  readonly width?: number;
  /** Min width, in pixels, of the cell */
  readonly minWidth?: number;
  /** Max width, in pixels, of the cell */
  readonly maxWidth?: number;
  /** True if column is initially hidden
   * @Notice To show/hide columns after the first render use PAGridAPI.setColumnVisible()
   */
  readonly hide?: boolean;
  /** Column is primary field */
  readonly isPrimary: boolean;
}

export type ColumnDataType =
  | 'Text'
  | 'Email'
  | 'Phone'
  | 'Ticker'
  | 'URL'
  | 'TextArea'
  | 'Lookup'
  | 'Customer'
  | 'Owner'
  | 'MultiSelectPicklist'
  | 'OptionSet'
  | 'TwoOptions'
  | 'Duration'
  | 'Language'
  | 'Multiple'
  | 'TimeZone'
  | 'Integer'
  | 'Currency'
  | 'Decimal'
  | 'FloatingPoint'
  | 'AutoNumber'
  | 'DateOnly'
  | 'DateAndTime'
  | 'Image'
  | 'File'
  | 'Persona'
  | 'RichText'
  | 'UniqueIdentifier';